requirements.txt
pyirsdk == 1.3.2
PyYAML == 6.0.2
flask==3.0.3
Werkzeug==3.0.4
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2
colorama==0.4.6
MarkupSafe==3.0.1